test(e2e): expand cooking-animation coverage and rename file
The previous e2e tests covered only the reasoning-gap slice of the
cooking animation. Two equally-common user-visible states had no
coverage:

  - Tool execution restarts the spinner (#6, #7): add a test that
    scripts a bash tool call with a 'sleep 1' command. The fake
    server's first response emits the tool call, the binary executes
    it (1s), then the fake server's second response produces the
    final text. Asserts the cooking frame is visible during the tool
    and gone after.

  - 'Worked for Xs' message after turn_end (#8): submit a prompt,
    wait for the response text, wait for /Worked for/. The 2.5s
    auto-hide timer is timer-dependent and fragile in a real-terminal
    e2e, so it's not asserted here (the unit test covers it precisely
    with fake timers).

Rename indicator-thinking.test.ts -> cooking-animation.test.ts
because the file now covers the full cooking-animation lifecycle,
not just the thinking slice. Updated artifactName values to match.

Extracted a small runSession() helper that accepts an optional
useThinkingModel flag, so each test declares its own model
requirement inline instead of repeating the models:/extraArgs:
boilerplate.

All 8 e2e tests pass. Unit tests unchanged.
